fix untracked files

+++ jeq.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python3
 import argparse
 import json
+from operator import eq, lt, gt, ne
+import re
 import sys
+import itertools
 
 __author__ = 'darryl'
 
@@ -16,9 +19,7 @@ def get_value(d, argval):
             if k == argval:
                 return d[k]
             if isinstance(v, list):
-                for i in v:
-                    print(get_value(i, argval))
-                return None
+                return get_value(v, argval)
             elif isinstance(v, dict):
                 if get_value(v, argval):
                     return get_value(v, argval)
@@ -35,9 +36,7 @@ def delete_key(d, argval):
                 del d[k]
                 return d
             elif isinstance(v, list):
-                for i in v:
-                    v[v.index(i)] = delete_key(i, argval)
-                return v
+                    return delete_key(v, argval)
             elif isinstance(v, dict):
                 if get_value(v, argval):
                     d[k] = delete_key(v, argval)
@@ -45,26 +44,44 @@ def delete_key(d, argval):
 
 
 def find_key(d, argval):
-    val_k, val_v = argval.split(':')
+    opts = {
+        '=': eq,
+        '<': lambda a, b: int(a) < int(b),
+        '>': lambda a, b: int(a) > int(b),
+        '!': ne
+    }
+    pattern = re.compile('(\w+)([=><!])(\w+)')
+    argvalre = re.match(pattern, argval)
+    val_k = argvalre.group(1)
+    val_v = argvalre.group(3)
+    opt = opts[argvalre.group(2)]
     if isinstance(d, list):
         for i in d:
             d[d.index(i)] = find_key(i, argval)
         return d
     else:
         for k, v in d.items():
-            if k == val_k and v == val_v:
+            if k == val_k and opt(v, val_k):
                 return d
             elif isinstance(v, list):
-                for i in v:
-                    v[v.index(i)] = find_key(i, argval)
-                return v
+                    return find_key(v, argval)
             elif isinstance(v, dict):
                 if get_value(v, val_v):
                     return find_key(v, argval)
 
 
+def parse_index(idx_str):
+    try:
+        return int(idx_str)
+    except ValueError:
+        r1, r2 = idx_str.split(':')
+        return [int(i) for i in range(int(r1), int(r2)+1)]
+
+
 def get_index(d, argval):
-    pass
+    f = set(itertools.chain([parse_index(i) for i in argval.split(',')]))
+    if isinstance(d, list):
+        return [i for i in d if d.index(i) in f]
 
 
 def main():
@@ -73,27 +90,24 @@ def main():
     parser = argparse.ArgumentParser(description="Pretty print and modify JSON files")
     parser.add_argument('-g', '--get', dest='get', action='store',
                         help='Display the value of a given key from all entries')
+    parser.add_argument('-f', '--find', dest='find', action='store',
+                        help='Only display entries where a key has a certain value '
+                             '(argformat: key[=<>!]value')
+    parser.add_argument('-n', '--index', dest='idx', action='store',
+                        help='Display entries with given index only (argformat: 1:10)')
     parser.add_argument('-d', '--delete', dest='delete', action='store',
                         help='Delete a given key and display the resulting output of all entries')
-    parser.add_argument('-f', '--find', dest='find', action='store',
-                        help='Only display entries where key has certain value')
-    parser.add_argument('-n', '--index', dest='idx', action='store',
-                        help='Display entries with given index only')
+    parser.add_argument('-m', '--merge', dest='merge', action='store',
+                        help='Merge output with other JSON file')
+    parser.add_argument()
     args = parser.parse_args()
 
-    if args.get:
-        res = get_value(res, args.get)
+    res = find_key(res, args.find) if args.find else res
+    res = delete_key(res, args.delete) if args.delete else res
+    res = get_index(res, args.idx) if args.idx else res
+    res = get_value(res, args.get) if args.get else res
 
-    if args.find:
-        res = find_key(res, args.find)
-
-    if args.idx:
-        res = get_index(res, args.idx)
-
-    if args.delete:
-        res = delete_key(res, args.delete)
-
-    if res:
+    if res and isinstance(res, dict):
         print(json.dumps(res, indent=4, separators=(',', ': ')))
